removed MgNode.enabled property, redundant with MgLayer.alpha
Removing it also avoids the thorny problem of how to animate that
boolean value. Layers are still culled from the renderer / layer tree
when alpha = 0, so we don't really lose anything.

One note: if we add CA-style background filters, it's not possible to
cull layers with opacity = 0, if they have certain kinds of background
filter / compositing filter / mask combinations, but we'll cross that
bridge if/when necessary.